+++ docs/changelog.md
@@ -70,6 +70,7 @@ This is release candidate 5 for the upcoming StepUp 4.0 release.
   This is useful for ensuring reproducible builds.
   See [Configuration files](reference/configuration.md) for details.
 - A resource usage report is shown ad the end of the file `.stepup/director.log`.
+- An SQL debug log option, to check query plans and execution times.
 
 ### Changed